split Solver base class
