+++ conda/plugins/solvers/classic/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (C) 2012 Anaconda, Inc
+# SPDX-License-Identifier: BSD-3-Clause
+"""The classic solver subpackage."""
+
+from .solver import ClassicSolver  # noqa
